+++ .eslintrc.js
@@ -46,7 +46,6 @@ module.exports = {
     'eol-last': 2,
     'eqeqeq': [2, 'allow-null'],
     'generator-star-spacing': [2, { 'before': true, 'after': true }],
-    'handle-callback-err': [2, '^(err|error)$' ],
     'indent': [2, 2, { 'SwitchCase': 1 }],
     'jsx-quotes': [2, 'prefer-double'],
     'key-spacing': [2, { 'beforeColon': false, 'afterColon': true }],