+++ docs/demos/views/coupon.vue
@@ -115,7 +115,7 @@ export default {
       });
     },
     randomId(max = 999999) {
-      return Math.floor(Math.random() * max) + 1
+      return Math.floor(Math.random() * max) + 1;
     }
   }
 };