+++ packages/button/index.vue
@@ -16,12 +16,7 @@
     ]"
     @click="onClick"
   >
-    <loading
-      v-if="loading"
-      class="van-button__icon-loading"
-      type="circle"
-      :color="type === 'default' ? 'black' : 'white'"
-    />
+    <loading v-if="loading" size="20px" :color="type === 'default' ? 'black' : 'white'" />
     <span class="van-button__text">
       <slot>{{ text }}</slot>
     </span>